Moved AccordionHeader component into AccordionHeader.js file
